requirements.txt
# FastMCP (Model Context Protocol) dependencies
fastmcp==2.12.3

# Google API dependencies
google-api-python-client==2.182.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-auth==2.40.3

# Groq AI dependencies
groq==0.31.1


# Utility dependencies
python-dotenv==1.1.1

# Development dependencies (optional)
pytest==8.4.2
pytest-asyncio==1.2.0
black==25.9.0
flake8==7.3.0